// src/shared/api/generated/model/auctionListItemMainAucType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Тип аукциона:
 *
 * * **Request** — заявочный (1)
 * * **Up** — на повышение (2)
 * * **Down** — на понижение (3)
 * * **FixPrice** — фиксированная цена (4)
 * * **Unknown** — неизвестный тип
 */
export type AuctionListItemMainAucType =
  (typeof AuctionListItemMainAucType)[keyof typeof AuctionListItemMainAucType];

export const AuctionListItemMainAucType = {
  Request: 'Request',
  Up: 'Up',
  Down: 'Down',
  FixPrice: 'FixPrice',
  Unknown: 'Unknown',
} as const;