// src/shared/api/generated/model/operationType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Тип операции маршрутной точки:
 * - `Loading` — погрузка (1)
 * - `Unloading` — выгрузка (2)
 * - `Unknown` — неизвестный тип
 */
export type OperationType = (typeof OperationType)[keyof typeof OperationType];

export const OperationType = {
  Loading: 'Loading',
  Unloading: 'Unloading',
  Unknown: 'Unknown',
} as const;