// src/shared/api/generated/model/paymentDelayType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Тип отсрочки платежа:
 * - `CalendarDays` — календарные дни (1)
 * - `WorkDays` — рабочие дни (2)
 * - `Unknown` — неизвестный тип
 * @nullable
 */
export type PaymentDelayType = (typeof PaymentDelayType)[keyof typeof PaymentDelayType] | null;

export const PaymentDelayType = {
  CalendarDays: 'CalendarDays',
  WorkDays: 'WorkDays',
  Unknown: 'Unknown',
} as const;