// src/shared/api/generated/model/tradingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Торговый статус пользователя в аукционе:
 * - `NotParticipating` — не участвует (1)
 * - `Leading` — лидирует (2)
 * - `Losing` — перебит (3)
 * - `Winner` — победитель (4)
 * - `Confirmed` — подтверждён (5)
 * - `Unknown` — неизвестный статус
 */
export type TradingStatus = (typeof TradingStatus)[keyof typeof TradingStatus];

export const TradingStatus = {
  NotParticipating: 'NotParticipating',
  Leading: 'Leading',
  Losing: 'Losing',
  OnPending: 'OnPending',
  Confirmed: 'Confirmed',
  ChoosingWinner: 'ChoosingWinner',
  Winner: 'Winner',
  Accepted: 'Accepted',
  Unknown: 'Unknown',
} as const;